WebsoketSessionsTable で remove する箇所にバグがあったのを修正した
